Issues #41 Add errors when assert operators.
When operators are evaluated, we can found some asserter false. This
patch add a management of errors in the asserter. Operators name and
value is added in errors stack to identify which operator fail.

Two methods are available to identify if we have errors and to get there
errors:
* hasErrors return a boolean which identify if some errors have been
raised.
* getErrors return an array with operator name as key and arguments as
values.
* resetErrors clean all errors stored.