Grammar: Can apply dimension calls on a function.
It was only possible to call dimensions on a variable, so:

    x.y[7].z(42)

but it was not possible to start from a function call instead of a
variable. This PR fixes this by introducing the `chain` rule that
aggregates the “call dimensions” for variable and function call.

The AST is the same but the model needs to be updated. Backward
compatibility is kept. The `Model\Operator` class extends the
`Model\Bag\Context` to be able to call attributes, indexes and calls. A
context is a variable, this is a different naming.